backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.4.0
google-generativeai==0.8.3
faiss-cpu==1.8.0
numpy==1.26.4
